README.md
I am very excited to be uploading my first ever Python project to GitHub and my first project on GitHub! This is a simple times table calculator I have created for my first assignment in OOP Python at WSU Tech. The raw code includes only one .py file that has everything in it. As this is a simple program, there is not much to see but it demonstrates an understanding of basic programming concepts in Python. I also took extra time to clean up the code and make it as readable as possible.